% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_numeric_by_category.R
\name{analyze_numeric_by_category}
\alias{analyze_numeric_by_category}
\title{Numerical Analysis by Grouping}
\usage{
analyze_numeric_by_category(df, categorical_col, numeric_cols)
}
\arguments{
\item{df}{A data set}

\item{categorical_col}{(character; column name) Categorical or Grouping column to compare}

\item{numeric_cols}{(character vector of column names) Other numerical columns to create summary table of the variable means copared with respect to the previous grouping column}
}
\value{
A summary table
}
\description{
Reports Group Comparison like t-test/Mann WHitney/ANOVA/Kruskal-Wallis to compare mean difference across a Grouping Variable.
}
