% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_category_by_category.R
\name{categoricalAnalysis_by_grouping}
\alias{categoricalAnalysis_by_grouping}
\title{Categorical Analysis}
\arguments{
\item{df}{A data set}

\item{categorical_col}{(character; column name) Categorical or Grouping column to compare}

\item{factor_columns}{(character vector of column names) Other categorical columns to create summary of the variables chosen with their distribution compared with respect to the previous grouping column}
}
\value{
A summary table
}
\description{
Reports Chi-sq/Fisher's Exact test to compare distribution of subjects across a Grouping Variable.
}
