% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check.missing}
\alias{check.missing}
\title{Check percent of missing values in data frame or matrix pre and post-processing}
\usage{
check.missing(data, post_data = NULL, verbose = FALSE)
}
\arguments{
\item{data}{data frame or matrix prior to processing}

\item{post_data}{data frame or matrix after processing}

\item{verbose}{indicate verbose warning and error messages}
}
\value{
a character vector specifying the percent of missing values after processing
}
\description{
check.missing takes a data frame and compares the percentage of missing values
pre and post-processing within one imputation function. If missing values completely
replaced by imputed values, the remaining percent missing should be 0, otherwise
the user will see a warning indicating imputation wasn't performed or the regressors
used in the model contained missing values.
}
\examples{
set.seed(123)
data <- data.frame(x1 = stats::rnorm(100),x2 = stats::rnorm(100),y = stats::rnorm(100))
data$x1[sample(1:100, 20)] <- NA
data$x2[sample(1:100, 15)] <- NA
data$y[sample(1:100, 10)] <- NA

check.missing(data, data, verbose = TRUE)

}
