% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{cmean.impute}
\alias{cmean.impute}
\title{Conditional Mean Imputation (CMI)}
\usage{
cmean.impute(
  data,
  family = "AUTO",
  robust = FALSE,
  char_to_factor = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a numeric matrix or data frame of at least 2 columns.}

\item{family}{the distribution family of your observations. The family arguments defaults
to 'AUTO'; and it will automatically select a distribution family (gaussian, binomial, multinomial) based on the type of
variable (numeric or factor). The distribution family dictates the regression model used (lm,glm, multinom).
However, the user can change the family argument to match his response variable distribution
and the function will adapt to this input by using the generalized linear model or beta regression.}

\item{robust}{logical indicated whether to use robust estimation methods or ignore them. If set to 'TRUE',
the function will make use of robust linear and generalized linear models to make its prediction.}

\item{char_to_factor}{transform character variable to unordered factor variable}

\item{verbose}{verbose error handling}
}
\value{
a matrix or data frame containing the imputed dataset.
}
\description{
This method replaces missing values with the expected value of the missing variable, given
other variables in the dataset. Predictions of conditional mean performed using specific
regression models.
CMI is a univariate imputation method that leverages the relationship between
variables in the data to make informed predictions about missing values. Although
this method reduces bias and aims to maintain the relationship among variables, it
will yield residuals with less variation than the original data.

CMI will fail if a regressor also contains missing values, thus making imputation of
the target's missing values impossible. The user will want to understand and fill in
as much of the missing data as possible before imputing through this method. If the
regressors contain missing values, it is better to use multiple imputation techniques.
}
\examples{
set.seed(123)
data <- data.frame(x1 = c(stats::rnorm(87),rep(NA,13)),
x2 = stats::rnorm(100),y = stats::rnorm(100))
cmi_data <- cmean.impute(data)
summary(cmi_data)
}
