% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_and_bland_altman.R
\name{Corr_and_BlandAltman}
\alias{Corr_and_BlandAltman}
\title{Correlation & Bland Altman Plot}
\arguments{
\item{data}{A data set}

\item{v1}{(character; numeric column's name)}

\item{v2}{(character; numeric column's name)}

\item{var1}{(numeric vector)}

\item{var2}{(numeric vector)}
}
\value{
Correlation Summary with 95% CI & Bland Altman Plot
}
\description{
Calculates Correlation for 2 variables (eg: 2 methods of treatment) & produces a Bland Altman Plot
}
