% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_and_bland_altman.R
\name{correlation_test_and_blandaltmanplot}
\alias{correlation_test_and_blandaltmanplot}
\title{Correlation & Bland Altman Plot}
\arguments{
\item{data}{A data set}

\item{v1}{(character; numeric column's name) To calculate correlation & conduct appropriate test for the same.}

\item{v2}{(character; numeric column's name) To calculate correlation & conduct appropriate test for the same.}

\item{var1}{(numeric vector) To plot the Bland Altman plot by calculating the means and differences of the 2 variables chosen.}

\item{var2}{(numeric vector) To plot the Bland Altman plot by calculating the means and differences of the 2 variables chosen.}
}
\value{
Correlation Summary with 95\% CI (use function correlation_test) & Bland Altman Plot (use function blandaltmanplot)
}
\description{
Calculates Correlation for 2 variables (eg: 2 methods of treatment) & produces a Bland Altman Plot
}
