% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Descriptive statistics}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{data frame}

\item{...}{other arguments passed to/from other methods}
}
\value{
Data frame with summaries
}
\description{
Descriptive statistics
}
\examples{
mtcars \%>\% group_by(cyl) \%>\% describe()

}

