% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_functions.R
\name{explore_missingness_single}
\alias{explore_missingness_single}
\title{Function to create an aggr object for a single tibble or data frame}
\usage{
explore_missingness_single(dat, plot = TRUE)
}
\arguments{
\item{dat}{tibble or data frame}

\item{plot}{a logical indicating whether hte reslts should be plotted}
}
\value{
an aggr object. if plot=T plots missingness patterns and barplot of missingness proportions.
}
\description{
Function to create an aggr object for a single tibble or data frame
}
