% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{format_dates}
\alias{format_dates}
\title{Removes time stamp from date variables in a list of tibbles or data frame for better readability}
\usage{
format_dates(df_list)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}
}
\value{
a list of tibbles with date variables in %m-%d-%y format
}
\description{
Removes time stamp from date variables in a list of tibbles or data frame for better readability
}
