% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{format_dates_single}
\alias{format_dates_single}
\title{Removes time stamp from date variables in a data frame or tibble for better readability}
\usage{
format_dates_single(df)
}
\arguments{
\item{df}{a tibble or data frame}
}
\value{
a tibble with date variables in %m-%d-%y format
}
\description{
Removes time stamp from date variables in a data frame or tibble for better readability
}
