% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{gen.mar}
\alias{gen.mar}
\title{Generate correlated, synthetic normal variables with user-specified probability of MAR}
\usage{
gen.mar(len, rho, sigma, n_vars, na_prob)
}
\description{
Generate correlated, synthetic normal variables with user-specified probability of MAR
}
