% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{gen.mcar}
\alias{gen.mcar}
\title{Generate correlated, synthetic normal variables with user-specified probability of MCAR.}
\usage{
gen.mcar(len, rho, sigma, n_vars, na_prob = 0)
}
\description{
Generate correlated, synthetic normal variables with user-specified probability of MCAR.
}
