% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{gen.mnar}
\alias{gen.mnar}
\title{Generate a data frame with values missing
not at random}
\usage{
gen.mnar(len, rho, sigma, n_vars, na_prob = 0.1)
}
\arguments{
\item{len}{number of rows per column}

\item{rho}{desired correlation coefficient of generated variables. The length of rho
must be equal to the product of \code{n_vars} and half of \code{n_vars} minus one.}

\item{sigma}{desired standard deviation for each generated variable.}

\item{n_vars}{total number of variables to be generated. At least two variables must be provided.}

\item{na_prob}{desired probability of missingness in each variable set to 10\% by default.}
}
\value{
a data frame of at least 2 columns
}
\description{
Generate a dataset containing values missing not at random.
Specify the column length, correlation coefficient, standard deviation, number of columns
and desired probability of missing values to obtain a data frame of correlated observations
with missing values.
}
\details{
The MNAR mechanism requires missingness to be related to events that are not measured or observed.
This type of missingness cannot be determined by statistical analysis and will
produce biased estimates. \code{gen.mnar} uses a self-selection mechanism to create values missing
not at random.Others methods, such as logistic regression models, and exponential decay models
may be suitable to generate missing values but the self-selection  mechanism provides a
straight-forward implementation of the MNAR mechanism.
}
\examples{
syn_na <- gen.mnar(50,c(.25,.75,.044),c(1.1,.56,1.56),3,.15)
summary(syn_na)
}
