% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.mad}
\alias{geom.mad}
\title{Calculate the median of distances to the geometric mean}
\usage{
geom.mad(x, iters = 100, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector or matrix containing a set of points in n-dimensions}

\item{iters}{number of iterations used in Weiszfeld algorithm}

\item{na.rm}{set to FALSE by default to ignore NA. Set to TRUE to remove NA}

\item{...}{optional arguments to specify alternative tolerance for convergence
current tolerance set to 1e-8.}
}
\value{
a numeric vector of length 1L.
}
\description{
The median absolute deviation is a measure of statistical dispersion that is
more resilient to outliers than the standard deviation. Because the median absolute
deviation is more robust to outliers, it may prove useful in distributions with
mean or variance.
}
\examples{
d <- rlnorm(60,1,1.4)
geom.mad(d,iters = 50)

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(geom.mad, as.data.frame(x))
}
