% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.median}
\alias{geom.median}
\title{Calculate the geometric median for non-collinear points in 2D+ Euclidean Space}
\usage{
geom.median(x, iters = 1000, tol = 1e-08, na.rm = FALSE)
}
\arguments{
\item{x}{vector of non-collinear points}

\item{iters}{number of iterations used in Weiszfeld algorithm}

\item{tol}{tolerance value used to check convergence of changes between
successive guesses.Default set to 1e-8.}

\item{na.rm}{set to FALSE. To handle NAs, set to TRUE.}
}
\value{
a numeric vector of length 1L.
}
\description{
Using the Weiszfeld algorithm an iterative method to compute the geometric
median of a set of points in Euclidean space. In general, measures of central
tendency minimize the sum of the Euclidean distance from the center to each
point in the set. We try to come up with an
estimate of this center by first using our best guess and then approximating
it by calculating  the distance from each point to a newly chosen point in
the set and updating the center if the distance is smaller than that
calculated using our best, initial guess.
}
\details{
If your sample is a 1D set of points use \code{median} or \code{grouped.median} instead.
The \code{geom.median} function is used for samples of points in +2D space.
}
\examples{
x <- matrix(rlnorm(60,1,1.4),ncol = 4)
geom.median(x,iters = 500, tol = 1e-10)
}
