% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.sd}
\alias{geom.sd}
\title{Calculate the geometric standard deviation (GSD) of the geometric mean.}
\usage{
geom.sd(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{...}{additional parameters to be passed to 'geom.mean()'}

\item{na.rm}{indicate whether or not to remove NAs}
}
\value{
a numeric vector of length equal or greater than 1L.
}
\description{
The GSD is a dimensionless (unitless) multiplicative factor. It describes
the range from the mean divided by GSD to the mean multiplied by GSD. Because of
its multiplicative nature, the GSD cannot be added/subtracted from the mean.
}
\examples{
x <- matrix(rlnorm(60,1,1.4), ncol = 4)
geom.sd(x)
}
