% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.zscore}
\alias{geom.zscore}
\title{Calculate the geometric standardized score}
\usage{
geom.zscore(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{...}{additional parameters to be passed to 'geom.mean()'}

\item{na.rm}{indicate whether or not to remove NAs}
}
\value{
a matrix or array of z-scores of length > 1L.
}
\description{
Compute the geometric z score of each strictly positive value in the sample,
relative to the geometric mean and standard deviation typically used in log-normally
distributed populations.
}
\details{
The formula here makes use of the change of base property of logarithms.
}
\examples{
d <- rlnorm(60,1,1.4)
geom.zscore(d)

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(geom.zscore, as.data.frame(x))
}
