% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_basic_summaries}
\alias{get_basic_summaries}
\title{Function to get summary statistics for numeric or date variables from a list of tibbles or
data frames}
\usage{
get_basic_summaries(df_list, varname, data_type = is.numeric, digits = 2)
}
\arguments{
\item{df_list}{a list of tibbles or data frame}

\item{varname}{grouping variables (generally appl_year)}
}
\value{
a list of lists each consisting of a data frame of summary statistics
}
\description{
Function to get summary statistics for numeric or date variables from a list of tibbles or
data frames
}
