% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_basic_summaries_single}
\alias{get_basic_summaries_single}
\title{Function to easily summarize numeric or date variables in a tibble or dataframe 
grouped by another variable (generally application year)}
\usage{
get_basic_summaries_single(dat, varname, data_type = is.numeric, digits = 3)
}
\arguments{
\item{dat}{a tibble or data frame}

\item{varname}{variable to group by (generally appl_year)}

\item{data_type}{class of variables to summarize. should be is.numeric or is.POSIXct}
}
\value{
a list containing summary statistics for each variable stored in a df
}
\description{
Function to easily summarize numeric or date variables in a tibble or dataframe 
grouped by another variable (generally application year)
}
