% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{get_binary}
\alias{get_binary}
\title{Get names of columns that are binary (have two distinct values, excluding NAs) for a data frame}
\usage{
get_binary(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
a character vector containing column names
}
\description{
Get names of columns that are binary (have two distinct values, excluding NAs) for a data frame
}
