% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{get_binary_num_cols}
\alias{get_binary_num_cols}
\title{Get the binary numeric columns in a data frame plus appl_year and study_id}
\usage{
get_binary_num_cols(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
tibble containing binary numeric variables
}
\description{
Get the binary numeric columns in a data frame plus appl_year and study_id
}
