% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_binary_tables}
\alias{get_binary_tables}
\title{Tables of binary variables by year}
\usage{
get_binary_tables(dat, varname)
}
\arguments{
\item{dat}{a data frame or tibble}

\item{varname}{a grouping variable}
}
\value{
a list of tables
}
\description{
Tables of binary variables by year
}
