% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_functions.R
\name{get_cat_tables}
\alias{get_cat_tables}
\title{Displays a table showing the proportion of observations in the N largest 
categories for each categorical variable in a list of tibbles or data frames}
\usage{
get_cat_tables(df_list, varname, digits = 3, n = 10)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}

\item{varname}{grouping variable (generally appl_year)}

\item{digits}{number of digits for rounding tables}
}
\value{
a list of lists, each containing a table
}
\description{
Displays a table showing the proportion of observations in the N largest 
categories for each categorical variable in a list of tibbles or data frames
}
