% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_functions.R
\name{get_cat_tables_single}
\alias{get_cat_tables_single}
\title{Displays a table showing the proportion of observations in the N largest categories for each categorical variable in 
a tibble or data frame}
\usage{
get_cat_tables_single(dat, varname, digits = 3, n = 10)
}
\arguments{
\item{dat}{tibble or data frame}

\item{varname}{grouping variable (generally appl_year)}

\item{digits}{number of digits for rounding the tables. default is 3.}

\item{n}{number of distinct categories to dispaly. default is 10.}
}
\value{
a list of tables
}
\description{
Displays a table showing the proportion of observations in the N largest categories for each categorical variable in 
a tibble or data frame
}
