% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_functions.R
\name{get_complete_cases}
\alias{get_complete_cases}
\title{Function to get the proportion of complete observations for each tibble in a
list of tibbles or data frames}
\usage{
get_complete_cases(df_list, varname)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}

\item{varname}{an optional grouping variable}
}
\value{
a list of data frames
}
\description{
Function to get the proportion of complete observations for each tibble in a
list of tibbles or data frames
}
