% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_functions.R
\name{get_complete_cases_single}
\alias{get_complete_cases_single}
\title{Function to get the proportion of oberservations with complete data in a tibble}
\usage{
get_complete_cases_single(dat, varname)
}
\arguments{
\item{dat}{a tibble or data frame}

\item{varname}{an optional grouping variable}
}
\value{
a data frame. if varname is not specified data frame contains 1 observation. otherwise a two
column data frame is returned.
}
\description{
Function to get the proportion of oberservations with complete data in a tibble
}
