% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_cor_mat}
\alias{get_cor_mat}
\title{Function to get correlation matrices for all numeric variables in a list of tibbles (or data frames)}
\usage{
get_cor_mat(df_list, use = "pairwise.complete.obs", round_digits = 3)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}

\item{use}{can be any of the character strings from the use argument in cor function}

\item{round_digits}{number of digits used for rounding correlation matrices}
}
\value{
a list of correlation matrices
}
\description{
Function to get correlation matrices for all numeric variables in a list of tibbles (or data frames)
}
