% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_cor_mat_by_year}
\alias{get_cor_mat_by_year}
\title{Get correlation matrices by year for a list of tibbles or data frames}
\usage{
get_cor_mat_by_year(df_list, varname)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}
}
\value{
a list of correlation matrices
}
\description{
Get correlation matrices by year for a list of tibbles or data frames
}
