% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_cor_mat_single}
\alias{get_cor_mat_single}
\title{Get correlation matrix for numeric variables in a data frame or tibble}
\usage{
get_cor_mat_single(dat, use = "pairwise.complete.obs", round_digits = 3)
}
\arguments{
\item{dat}{a data frame or tibble}

\item{use}{a character string giving a method for computing covariances in the presence of missing values}

\item{round_digits}{number of digits used for rounding correlation matrix}
}
\value{
correlation matrix for a single tibble
}
\description{
Get correlation matrix for numeric variables in a data frame or tibble
}
