% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_functions.R
\name{get_ndistinct}
\alias{get_ndistinct}
\title{Function to get the number of distinct categories for each categorical variable in a list of tibbles or data frames}
\usage{
get_ndistinct(df_list, varname)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}

\item{varname}{an optional grouping variable (generally appl_year)}
}
\value{
a list of tibbles containing the number of distinct categories for each categorical variable
 within each tibble from df_list
}
\description{
Function to get the number of distinct categories for each categorical variable in a list of tibbles or data frames
}
