% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_functions.R
\name{get_ndistinct_single}
\alias{get_ndistinct_single}
\title{Function to get the number of categories for each categorical variable in a tibble or data frame}
\usage{
get_ndistinct_single(dat, varname)
}
\arguments{
\item{dat}{a tibble or data frame}

\item{varname}{an optional grouping variable (generally appl_year)}
}
\value{
a tibble with number of distinct categories in each categorical variable
}
\description{
Function to get the number of categories for each categorical variable in a tibble or data frame
}
