% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{get_var_types}
\alias{get_var_types}
\title{Gets variable type information for each tibble in a list of tibbles}
\usage{
get_var_types(df_list)
}
\arguments{
\item{df_list}{a list of data frames or tibbles}
}
\value{
a list of data frames, each containing two columns
}
\description{
Gets variable type information for each tibble in a list of tibbles
}
