% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{get_var_types_single}
\alias{get_var_types_single}
\title{Prints a table containing the class of all variables in the data set along
with whether the variable is binary or not. In this case, binary means that 
there are at most two distinct values for the variable, ignoring NA.}
\usage{
get_var_types_single(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
a data frame with two columns
}
\description{
Prints a table containing the class of all variables in the data set along
with whether the variable is binary or not. In this case, binary means that 
there are at most two distinct values for the variable, ignoring NA.
}
