% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{get_yesno}
\alias{get_yesno}
\title{Get names of columns that only have Yes, No, or NA responses}
\usage{
get_yesno(dat)
}
\arguments{
\item{dat}{a data frame or tibble}
}
\value{
character vector containing column names
}
\description{
Get names of columns that only have Yes, No, or NA responses
}
