% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_central.R
\name{harm.mean}
\alias{harm.mean}
\title{Calculate the harmonic mean for matrices, data frames and vectors.}
\usage{
harm.mean(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{an object containing a set of observations in a matrix, data frame or vector.}

\item{na.rm}{logical by default set to FALSE. To remove NAs set argument to TRUE}

\item{...}{additional arguments passed to or from other methods}
}
\value{
a numeric vector of length equal or greater than 1L
}
\description{
The harmonic mean is the reciprocal of the arithmetic mean of the
reciprocals of a set of observations. It is a more robust measure of central
tendency than the arithmetic mean in the presence of outliers since it tends
to the smallest elements of a set of observations.
}
\examples{
d <- rlnorm(60,1,1.4)
harm.mean(d)

x <- matrix(rlnorm(60,1,1.4),ncol = 4)
sapply(x,harm.mean)
}
