% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_functions.R
\name{initialize_cat_tables}
\alias{initialize_cat_tables}
\title{Helper function for get_cat_tables_single to get a list of tables for all categorical variables}
\usage{
initialize_cat_tables(dat, varname, digits = 3)
}
\arguments{
\item{dat}{a tibble or dataframe}

\item{varname}{grouping variable (generally appl_year)}

\item{digits}{number of digits for rounding proportions}
}
\value{
a list of tables each containing the proportion of observations in each category grouped
 by the grouping variable
}
\description{
Helper function for get_cat_tables_single to get a list of tables for all categorical variables
}
