% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM summary table.R
\name{km_summary}
\alias{km_summary}
\title{Kaplan Meier Survival Analysis Table (Standard Format)}
\arguments{
\item{data}{A data set}

\item{time_var}{(column name of time duration) A numeric column containing the duration of survival of all subjects.}

\item{event_var}{(column name of status/event variable) A binary numeric column containing the indicators whether the concerned event has occurred or not.}

\item{stratify_var}{(column name of stratification variable) A binary column containing the names of groups in which the study subjects belongs to.}

\item{duration_unit}{Pass on the choice of duration unit. Choose from "days", "weeks", "months", and "years". Default: "days".}
}
\value{
A standard format dataframe from the Kaplan Meier Survival Analysis. Reports the Median Follow-up of the study, the Variable, its Levels, total subjects in each level, no. of events in each level, Mediabn Survival (95\% CI), and p-value from Log-Rank test.
}
\description{
Reports a summary table from the KM analysis, with p-value (log rank test), median follow-up (using reverse KM) & median survival (with 95\% CI).
}
