% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate kurtosis}
\usage{
kurtosis(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of length +2L}

\item{...}{additional arguments to be passed to function}

\item{na.rm}{option to remove or keep NAs, set to FALSE by default.}
}
\value{
a numeric vector of length 1
}
\description{
Calculate kurtosis
}
\examples{
s <- rnorm(30,1,1.4)
kurtosis(s)

air_data <- datasets::airquality
mapply(kurtosis,air_data)
}
