% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_logistic_regression.R
\name{logistic_regression_summary}
\alias{logistic_regression_summary}
\title{Logistic Regression Summary}
\arguments{
\item{data}{A data set}

\item{response_var}{(character; binary column's name)}

\item{predictor_vars}{(character vector; covariate columns' names)}

\item{response_ref}{(reference level of the binary outcome)}

\item{predictor_refs}{(named list of references of each categorical covariate)}
}
\value{
A Logistic Model summary with ODDS RATIO (95% CI), p-value & Model Performance
}
\description{
Creates a Logistic Regression Model; Preferred: categorical columns are formatter as factors with sub levels.
}
