% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{mar.transform}
\alias{mar.transform}
\title{Transform  data to missing values at random.}
\usage{
mar.transform(input, target, features, na_rate)
}
\arguments{
\item{input}{data to transform using MAR mechanism}

\item{target}{variable to be used as causative feature}

\item{features}{variables to which NA values are introduced using a causative feature}

\item{na_rate}{proportion of missing values to be added to data}
}
\value{
a matrix or data frame containing NAs
}
\description{
Transform a complete case dataset according to the MAR mechanism.The MAR
mechanism assumes that the probability of missingness in a variable depends
on the observed data but not on the missing data itself. This function introduces
missing values in selected features of a dataset, with the missing values determined
by the values of the causative (target) feature.
}
\examples{
set.seed(123)
data <- gen.mcar(100,rho = c(.15,.25,.12,.45,.34,.54),sigma = c(1,2,1,2),n_vars = 4, na_prob = 0)

mar.transform(data,"V1",c("V2","V3"), na_rate = .25)

}
