% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{mcar.transform}
\alias{mcar.transform}
\title{Transform non-missing data to missing values with user-specified probability of MCAR}
\usage{
mcar.transform(input, na_prob)
}
\arguments{
\item{input}{data to transform using MCAR mechanism}

\item{na_prob}{desired probability of NA count defined as the probability of success
in a Bernoulli trial.}
}
\value{
a data frame or matrix containing NAs
}
\description{
Transform a complete case dataset according to the MCAR mechanism.The MCAR
mechanism assumes that the probability of missingness is the same for all cases,
and therefore,the missing data are not related to the observed data. \code{transform.mcar}
uses the binomial distribution to generate NA  an index of possible NA values that will
replace a percentage of data points in the input.
}
\examples{
set.seed(123)
data <- data.frame(x1 = stats::rnorm(100),x2 = stats::rnorm(100),y = stats::rnorm(100))
mcar.transform(data,na_prob = .25)

}
