% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_missing.R
\name{mnar.transform}
\alias{mnar.transform}
\title{Transform data to missing values not at random}
\usage{
mnar.transform(input, target, features, na_rate)
}
\arguments{
\item{input}{data to transform using MAR mechanism}

\item{target}{variable to be used as causative feature}

\item{features}{variables to which NA values are introduced using a causative feature}

\item{na_rate}{proportion of missing values to be added to data}
}
\value{
a matrix or data frame containing NAs
}
\description{
The MNAR mechanism assumes that the probability and cause of missingness are unknown to us and lie
in the unobserved data. Under this mechanism we generate missing values by looking at each
variable's lowest values and replace them with NA.
}
\examples{
set.seed(123)
data <- gen.mcar(100,rho = c(.15,.25,.12,.45,.34,.54),sigma = c(1,2,1,2),n_vars = 4, na_prob = 0)

mnar.transform(data,"V1",c("V2","V3"), na_rate = .25)
}
