% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_logistic_regression.R
\name{mv_logistic_regression_summary}
\alias{mv_logistic_regression_summary}
\title{Multivariate Logistic Regression Summary}
\arguments{
\item{data}{A data set}

\item{response_var}{(character; binary column's name)}

\item{predictor_vars}{(character vector; covariate columns' names)}

\item{response_ref}{(reference level of the binary outcome)}

\item{predictor_refs}{(named list of references of each categorical covariate)}
}
\value{
A Logistic Model summary with ODDS RATIO (95% CI), p-value & Model Performance. Returns the summary by building the model with outcome variable & all the predictor variables.
}
\description{
Creates a Multivariate Logistic Regression Model summary table;   Preferred: categorical columns are formatted as factors with sub levels.
}
