% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_functions.R
\name{pct_miss_table}
\alias{pct_miss_table}
\title{Function to create a matrix with the percent of observations missing in each variable 
in a list of data frames or tibbles grouped by another variable (generally appl_year)}
\usage{
pct_miss_table(df_list, varname, round_digits = 1)
}
\arguments{
\item{varname}{grouping variable}

\item{tbl_list}{list of tibbles or data frames}

\item{round}{number of digits for rounding}
}
\value{
a list of matrices
}
\description{
Function to create a matrix with the percent of observations missing in each variable 
in a list of data frames or tibbles grouped by another variable (generally appl_year)
}
