% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_functions.R
\name{pct_miss_table_single}
\alias{pct_miss_table_single}
\title{Function to create a matrix with the percent of observations missing in each variable
in a data frame or tibble grouped by another variable (generally appl_year)}
\usage{
pct_miss_table_single(dat, varname, round_digits = 1)
}
\arguments{
\item{dat}{tibble or data frame}

\item{varname}{grouping variable}

\item{round_digits}{number of digits for rounding}
}
\value{
a matrix
}
\description{
Function to create a matrix with the percent of observations missing in each variable
in a data frame or tibble grouped by another variable (generally appl_year)
}
