% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{Plot}
\alias{Plot}
\title{Scatter/Box/Bar Plot}
\arguments{
\item{df}{A data set}

\item{type}{(character; choose from "scatter" or "boxplot" or "barplot")}

\item{var1}{(character; column name)}

\item{var2}{(character; column name)}
}
\value{
PLOTS
}
\description{
Plots using data, name of plot, & column names
}
