% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_central.R
\name{quad.mean}
\alias{quad.mean}
\title{Calculate the quadratic mean}
\usage{
quad.mean(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of observations}

\item{...}{additional arguments passed to or from other methods}

\item{na.rm}{logical by default set to FALSE. To remove NAs set argument to TRUE}
}
\value{
a numeric vector of length 1L
}
\description{
Calculate the absolute magnitude for a numeric vector where greater values
are given more importance than lesser values. The arithmetic mean is typically
equal or greater than the arithmetic mean.
}
\details{
The quadratic mean or root mean square is scale invariant but not shift invariant, shifting
the input values by a fix amount will not change the point estimate.
}
\examples{
d <- rlnorm(60,1,1.4)
quad.mean(d)

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(quad.mean, as.data.frame(x))

}
