% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{recode_yesno}
\alias{recode_yesno}
\title{Recodes Y/N/NA vars to 1/0/NA and converts to numeric for a list of tibbles}
\usage{
recode_yesno(df_list)
}
\arguments{
\item{df_list}{a list of tibbles or data frames}
}
\value{
A list of tibbles with Yes/No variables recoded to 1/0
}
\description{
Recodes Y/N/NA vars to 1/0/NA and converts to numeric for a list of tibbles
}
