% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_functions.R
\name{recode_yesno_single}
\alias{recode_yesno_single}
\title{Recodes Y/N/NA vars to 1/0/NA and converts to numeric for a single tibble/data frame}
\usage{
recode_yesno_single(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
a tibble or data frame containing only recoded columns
}
\description{
Recodes Y/N/NA vars to 1/0/NA and converts to numeric for a single tibble/data frame
}
