% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{seventyfive_quantile}
\alias{seventyfive_quantile}
\title{Helper function to return the third quartile within summarize functions}
\usage{
seventyfive_quantile(a_vec, ...)
}
\arguments{
\item{a_vec}{}

\item{...}{}
}
\value{
value of third quartile
}
\description{
Helper function to return the third quartile within summarize functions
}
