% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{stoc.impute}
\alias{stoc.impute}
\title{Stochastic regression imputation (SRI)}
\usage{
stoc.impute(
  data,
  family = "AUTO",
  tol = NULL,
  robust = FALSE,
  char_to_factor = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a numeric matrix or data frame of at least 2 columns.}

\item{family}{the distribution family of your observations. The family arguments defaults
to 'AUTO'; and it will automatically select a distribution family (gaussian, binomial, multinomial) based on the type of
variable (numeric or factor). The distribution family dictates the regression model used (lm,glm, multinom).
However, the user can change the family argument to match his response variable distribution
and the function will adapt to this input by using the generalized linear model or beta regression.}

\item{tol}{tolerance,a numeric vector of length 1 used as multiplicative factor to standard deviation for generalized linear models.
As the sample size increases, the tolerance value should be decreased to represent the decreasing variability of the sample estimate.}

\item{robust}{logical indicated whether to use robust estimation methods or ignore them. If set to 'TRUE',
the function will make use of robust linear and generalized linear models to make its prediction.}

\item{char_to_factor}{transform character variable to unordered factor variable}

\item{verbose}{verbose error handling}
}
\value{
a matrix or data frame containing the imputed dataset.
}
\description{
This method corrects the lack of variability in conditional mean imputation (CMI) by adding
an error term to the conditional mean calculation. SRI is more effective than CMI in reducing
bias in the imputed values and works well with MCAR and MAR data.
}
\examples{
data <- data.frame(x1 = c(stats::rnorm(87),rep(NA,13)),
x2 = stats::rnorm(100),y = stats::rnorm(100))
sri_data <- stoc.impute(data,tol = 1e-3)
summary(sri_data)
}
