% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfy.R
\name{tidyfy}
\alias{tidyfy}
\alias{tidyfy.data.frame}
\alias{tidyfy.default}
\alias{tidyfy.grouped_df}
\title{Tidy data frame}
\usage{
tidyfy(x, ...)

\method{tidyfy}{grouped_df}(x, ...)

\method{tidyfy}{data.frame}(x, by, ...)

\method{tidyfy}{default}(x, ...)
}
\arguments{
\item{x}{R object, usually a data frame}

\item{...}{other arguments passed to/from other methods}

\item{by}{character, variables names to split descriptives by}
}
\value{
List of data frames
}
\description{
Tidy data frame
}

