% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{twentyfive_quantile}
\alias{twentyfive_quantile}
\title{Helper function to return the first quartile within summarize functions}
\usage{
twentyfive_quantile(a_vec, ...)
}
\arguments{
\item{a_vec}{}

\item{...}{other arguments passed to quantile function}
}
\value{
value or first quartile
}
\description{
Helper function to return the first quartile within summarize functions
}
