% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_cox_ph.R
\name{uni_cox_ph_model_summary}
\alias{uni_cox_ph_model_summary}
\title{Univariate Cox Proportional Hazard Regression Summary}
\usage{
uni_cox_ph_model_summary(
  data,
  start_date_var,
  event_date_var,
  end_date_var,
  event_var,
  event_marker,
  predictor_vars,
  predictor_refs
)
}
\arguments{
\item{data}{A data set}

\item{start_date_var}{The start date of survival. Eg: Date of Registration/Randomization/Treatment start, etc.}

\item{event_date_var}{The date of occurrence of the Event of Interest Eg: Date of Death/Relapse/Progression/Recurrence, etc.}

\item{end_date_var}{The end date of survival. Eg: Date of Death/Last Follow Up (LFU), etc.}

\item{event_var}{binary variable: The variable that tells us if the Event of Interest has occurred or not. Usually, '1' for Event Occurred and '0' for No Event Occurrence.}

\item{event_marker}{single input: Choose the correct option among the 2 sub-categories of event_var. The event_var is likely to have binary options like "Dead" or "Alive", '1' or '0', "Progressed" or "Not Progressed", etc.}

\item{predictor_vars}{(character vector; covariate columns' names)}

\item{predictor_refs}{(named list of reference levels of each categorical covariate)}
}
\value{
A Logistic Model summary with ODDS RATIO (95\% CI), p-value, along with OR Forest Plot. Returns the summary by building the model with the outcome variable & one predictor variable at a time.
}
\description{
Creates a Univariate Cox Proportional Hazard Regression Model summary table;   Preferred: categorical columns are formatted as factors with sub levels.
}
