% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_logistic_regression.R
\name{univariate_logistic_regression_summary}
\alias{univariate_logistic_regression_summary}
\title{Univariate Logistic Regression Summary}
\arguments{
\item{data}{A data set}

\item{response_var}{(character; binary column's name)}

\item{predictor_vars}{(character vector; covariate columns' names)}

\item{response_ref}{(reference level of the binary outcome)}

\item{predictor_refs}{(named list of references of each categorical covariate)}
}
\value{
A Logistic Model summary with ODDS RATIO (95% CI), p-value & Model Performance. Returns the summary by building the model with outcome variable & exactly 1 predictor variable, at a time. For example, if 4 predictors are selected, a total of 4 models would be built, with each predictor entering once in a simple logistic model, Y ~ X. The summary henceforth presented would be of the 4 models built.
}
\description{
Creates multiple Single-Predictor Logistic Regression Model summary table;   Preferred: categorical columns are formatted as factors with sub levels.
}
