% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_classes.R
\name{variable_classes}
\alias{variable_classes}
\title{Columns of data frame by class}
\usage{
variable_classes(dat)
}
\arguments{
\item{dat}{data frame or object inheriting from one}
}
\value{
Returns a list with as many elements as there are classes of columns
in \code{dat}. Elements are named with the names of the classses. Each
element is a character vector of column names of \code{dat} of the
corresponding class.
}
\description{
Columns of data frame by class
}

