% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_central.R
\name{w.harm.mean}
\alias{w.harm.mean}
\title{Calculate the weighted harmonic mean for matrices, data frames and vectors}
\usage{
w.harm.mean(x, w = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{an object containing a set of observations whose weighted harmonic mean
is to be computed}

\item{w}{a numerical vector of weights the same length as x}

\item{na.rm}{logical by default set to FALSE. To remove NAs set argument to TRUE}

\item{...}{additional arguments passed to or from other methods}
}
\value{
a numeric vector or length equal or greater than 1L
}
\description{
Calculate the weighted harmonic mean for matrices, data frames and vectors
}
\examples{
x <- matrix(rlnorm(60,1,1.4),ncol = 4)
w <- matrix(rnorm(60,0,0.23), ncol = 4)
w.harm.mean(x,w)
}
