% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others_central.R
\name{weight.harm.mean}
\alias{weight.harm.mean}
\title{Calculate the weighted harmonic mean}
\usage{
weight.harm.mean(x, weights = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{an object containing a set of observations whose weighted harmonic mean
is to be computed}

\item{weights}{a numerical vector of weights the same length as x}

\item{na.rm}{logical by default set to FALSE. To remove NAs set argument to TRUE}

\item{...}{additional arguments passed to or from other methods}
}
\value{
a numeric vector or length equal or greater than 1L
}
\description{
Calculate the weighted harmonic mean
}
\examples{
d <- rlnorm(60,1,1.4)
w <- rlnorm(60,2,3.23)
weight.harm.mean(d,w)

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
w <- matrix(rlnorm(60,1,1), ncol = 4)
mapply(weight.harm.mean, as.data.frame(x),as.data.frame(w))
}
