% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-package.R
\docType{package}
\name{describe-package}
\alias{describe}
\alias{describe-package}
\title{describe: Compute varied descriptive statistics,create visualizations and handle missing data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

To provide estimates of central tendency for different distributions along with visualization options and imputation methods for handling missing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpmonteagudo28/describe}
  \item Report bugs at \url{https://github.com/jpmonteagudo28/describe/issues}
}

}
\author{
\strong{Maintainer}: JP Monteagudo \email{jpmonteagudo2014@gmail.com} (\href{https://orcid.org/0009-0003-6465-6658}{ORCID}) [copyright holder]

}
\keyword{internal}
