% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.var}
\alias{geom.var}
\title{Calculate the geometric variance}
\usage{
geom.var(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{...}{additional parameters to be passed to 'geom.mean()'}

\item{na.rm}{indicate whether or not to remove NAs}
}
\value{
a numeric vector of length equal or greater than 1L.
}
\description{
Spread of values around the geometric mean. The geometric standard deviation
can be described as a geometric factor and this cannot be added/subtracted from
the geometric mean.
}
\examples{
d <- rlnorm(60,1,1.3)
geom.var(d)

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(geom.var,as.data.frame(x))
}
