% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.var}
\alias{geom.var}
\title{Calculate the geometric variance of a set of observations of non-negative,
non-zero observations.}
\usage{
geom.var(x, impute_method = NULL, k = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{na.rm}{indicate whether or not to remove NAs}

\item{...}{additional parameters to be passed to 'geom.mean()'}
}
\value{
a numeric vector of length equal or greater than 1L.
}
\description{
Calculate the geometric variance of a set of observations of non-negative,
non-zero observations.
}
\examples{
x <- matrix(rlnorm(60,1,1.4), ncol = 4)
geom.var(x)
}
