% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_central.R
\name{geom.var}
\alias{geom.var}
\title{Calculate the geometric variance of a set of observations. The variance is also
a dimensionless factor.}
\usage{
geom.var(x, ..., na.rm = FALSE)
}
\arguments{
\item{...}{additional parameters to be passed to 'geom.mean()'}

\item{x:}{a numeric vector, matrix or data frame}

\item{na.rm:}{indicate whether or not to remove NAs}
}
\value{
a numeric vector of length equal or greater than 1L.
}
\description{
Calculate the geometric variance of a set of observations. The variance is also
a dimensionless factor.
}
\examples{
x <- matrix(rlnorm(60,1,1.4), ncol = 4)
geom.var(x)
}
