% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_central.R
\name{grouped.mean}
\alias{grouped.mean}
\title{Calculate mean for grouped data using frequency, either percent or counts
The function will split each character element in x and find the midpoint of
its lower and upper bounds.}
\usage{
grouped.mean(x, freq, freq_format = "count", na.rm = FALSE)
}
\arguments{
\item{x:}{character vector of categories or ranges}

\item{freq:}{numeric vector containing category counts or percentages}

\item{freq_format:}{intended freq format. Use percent if you want freq
converted to percent}

\item{na.rm:}{remove \code{NA} from calculations}
}
\value{
a numeric vector of length 1L
}
\description{
Calculate mean for grouped data using frequency, either percent or counts
The function will split each character element in x and find the midpoint of
its lower and upper bounds.
}
